Rewritten a lot of studd inside token/token.go
Divided SimpleToken and (Complex)Token

A SimpleToken is a Token that can be found by the tokenizer.
An example could be BoldToken

A (Complex)Token is a Token that is too
complex for be found by the tokenizer.

TextToken is a particular token, it is generated by the tokenizer
but contains metadata like Bold and Italic which are used by the parser

This version is not stable and contains a lot of bugs.

Some code is maintained for legacy and will be removed in a future commit